+++ build_web_compilers/lib/builders.dart
@@ -9,9 +9,9 @@ import 'package:build_web_compilers/build_web_compilers.dart';
 
 import 'package:path/path.dart' as p;
 
-Builder devCompilerBuilder(_) => DevCompilerBuilder();
+Builder devCompilerBuilder(_) => DevCompilerBuilder(useKernel: true);
 Builder webEntrypointBuilder(BuilderOptions options) =>
-    WebEntrypointBuilder.fromOptions(options);
+    WebEntrypointBuilder.fromOptions(options, useKernel: true);
 PostProcessBuilder dart2JsArchiveExtractor(BuilderOptions options) =>
     Dart2JsArchiveExtractor.fromOptions(options);
 PostProcessBuilder dartSourceCleanup(BuilderOptions options) =>