+++ build_web_compilers/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 2.0.0
+## 2.0.0-dev
 
+- Update to run DDC in kernel mode, and consume kernel outlines instead of
+  analyzer summaries.
 - Increased the upper bound for `package:analyzer` to `<0.37.0`.
 
 ## 1.2.0